PSR-2 cleanup.
