+++ src/Dersam/RT/Exceptions/HttpException.php
@@ -7,9 +7,9 @@
 
 namespace Dersam\RT\Exceptions;
 
-
 use Exception;
 
-class HttpException extends Exception{
+class HttpException extends Exception
+{
 
 }
