+++ src/Dersam/RT/Requests/V2Request.php
@@ -1,0 +1,29 @@
+<?php
+/**
+ * User: Sam
+ * Date: 4/15/2015
+ * Time: 7:38 PM
+ */
+
+namespace Dersam\RT\Requests;
+
+
+use Dersam\RT\Client;
+use Dersam\RT\Request;
+use Dersam\RT\Response;
+
+class V2Request extends Request{
+
+    /**
+     * @return Response
+     */
+    public function makeResponseInstance()
+    {
+        // TODO: Implement makeResponseInstance() method.
+    }
+
+    public function send(Client $client)
+    {
+        // TODO: Implement send() method.
+    }
+}
